Add EventHandlers module and move createEventHandlers there
