+++ swarm.cabal
@@ -664,6 +664,7 @@ library swarm-tui
   exposed-modules:
     Swarm.TUI.Border
     Swarm.TUI.Controller
+    Swarm.TUI.Controller.EventHandlers
     Swarm.TUI.Controller.EventHandlers.Frame
     Swarm.TUI.Controller.EventHandlers.Main
     Swarm.TUI.Controller.EventHandlers.REPL
